+++ embedding_generator.go
@@ -48,7 +48,7 @@ func (m *Model) EmbeddingGenerator(inputDims int, catDims []int, catIdxs []int, 
 			s := gorgonia.Must(gorgonia.Slice(x, nil, gorgonia.S(featInitIdx)))
 
 			if isCategorical {
-				s := gorgonia.Must(gorgonia.ConvType(s, tensor.Float64, tensor.Int))
+				s := gorgonia.Must(gorgonia.ConvType(s, tensor.Float32, tensor.Int))
 				cols[featInitIdx], _, err = embeddings[catFeatCounter](s)
 				if err != nil {
 					panic(err)
